Refactor to remove "incorrect loop" tracking.
We no longer check if a tile is on the loop incorrectly as well as valid or not. There are only valid or invalid tiles. This might bring back some fixxed bugs regarding loops as it touches the validation/revalidation process. Also refactored validation to use TileData instead of Hex.